part_5_production/chapter_19_deployment/requirements.txt
# requirements.txt
fastapi==0.124.0
uvicorn==0.38.0
openai==2.9.0
langchain-core==1.2.0
langchain-openai==1.1.1
python-dotenv==1.2.1
httpx==0.28.1
pydantic==2.12.5
numpy==2.3.5
langgraph==1.0.5